fix: moved shared host constants to a dependency-free file
