+++ src/utils/web-host.ts
@@ -1,16 +1,15 @@
 import { Platform } from 'react-native'
 import type { Platform as AnnouncementPlatform } from '@/__generated__/gql/graphql'
+import { NEULAND_DEV_HOST, NEULAND_WEB_HOST } from '@/utils/neuland-hosts'
 
 export type WebPlatform = 'web-dev' | 'web' | 'web-local'
 
-export const NEULAND_DEV_HOST = 'dev.neuland.app'
-export const NEULAND_WEB_HOST = 'web.neuland.app'
-
-export const NEULAND_APP_HOSTS = [NEULAND_DEV_HOST, NEULAND_WEB_HOST] as const
-
-export function isNeulandAppHost(hostname: string): boolean {
-	return hostname === NEULAND_DEV_HOST || hostname === NEULAND_WEB_HOST
-}
+export {
+	isNeulandAppHost,
+	NEULAND_APP_HOSTS,
+	NEULAND_DEV_HOST,
+	NEULAND_WEB_HOST
+} from '@/utils/neuland-hosts'
 
 export function resolveAnnouncementPlatform(
 	hostname?: string
